api/requirements.txt
aiocron==1.8
ampalibe==1.1.8
anyio==3.6.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
croniter==1.3.14
decorator==5.1.1
dnspython==2.3.0
fastapi==0.95.1
h11==0.14.0
idna==3.4
py==1.11.0
pydantic==1.10.7
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-toolbelt==0.10.1
retry==0.9.2
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tinydb==4.7.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.21.1
watchdog==3.0.0
